+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Next
 
 ### Added
+- Allow user to specify registry path without having to add semgrep.dev url
+  i.e.: instead of `--config https://semgrep.dev/p/r2c-ci` users can use `--config p/r2c-ci`
+- Allow user to specify snippet id withouth having to add semgrep.dev url
+  i.e.: instead of `--config https://semgrep.dev/s/username:snippetname`
+  users can use `--config username:snippetname`
 
 ### Fixed
 - More off by one fixes in autofix